requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.8.1+cu111
torchvision==0.9.1+cu111
torchaudio==0.8.1
opencv-python==4.5.3.56
pandas~=2.2.2
Pillow~=10.4.0
protobuf==3.17.3
pycm==3.2
pydicom==2.3.0
scikit-image~=0.24.0
scikit-learn==0.24.2
scipy~=1.13.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
timm==0.3.2
tqdm==4.62.1

numpy~=1.26.4
numba~=0.60.0